requirements.txt
streamlit
Pillow
transformers
requests
torch
torchvision
